pyproject.toml
[tool.poetry]
name = "paper_to_podcast"
version = "0.1.0"
description = "A pipeline that transforms research papers into two-person audio podcasts."
authors = ["Your Name <your_email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = ">=0.27.0"
requests = ">=2.31.0"
pydub = ">=0.25.1"
PyPDF2 = ">=3.0.1"
python-dotenv = ">=1.0.0"
langchain_chroma = "0.1.4"
langchain_community = "0.3.10"
langchain_core = "0.3.22"
langchain_openai = "0.2.11"
langchain_text_splitters = "0.3.2"
PyYAML = ">=6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
